Move delegation code to new module
